% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_IS_LM.R
\name{chart_IS_LM}
\alias{chart_IS_LM}
\title{Create a shiny app with IS-LM chart}
\usage{
chart_IS_LM()
}
\value{
It returns a interactive \link[shiny]{shinyApp}
}
\description{
Create a shiny app with IS-LM chart
}
\examples{
\dontrun{
chart_IS_LM()
}
}
